+++ postcss.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  plugins: {
+    'postcss-import': {options: { sourceMap: true }},
+    'postcss-preset-env': {}
+  }
+};